front: move train creation logic to its own hook
To make sure we don't break things in a way that is not relevant to our modal swapping, let's move the creation to its own hook that we will call from both the new and the old modal.
